refactor(groups): Extract GroupDetailScreen into modular components (#90, #102)
- Extract 5 widgets from GroupDetailScreen (2538 → 540 lines):
  - MembersBottomSheet, MemberOptionsSheet
  - DateRangePickerModal, CustomTimePickerModal
  - DayDetailSheet
- Fix inconsistent initialization pattern in FriendProvider and GroupProvider
  - Add shared _loadData() method used by both initialize() and refresh()
- Maintain backwards compatibility via re-export
